android: переключение подписок

+++ Spectator.Core/Model/ISnapshotCollectionModel.cs
@@ -7,6 +7,6 @@ namespace Spectator.Core.Model
 {
 	public interface ISnapshotCollectionModel
 	{
-		Task<IEnumerable<Snapshot>> GetAllAsync (int subscriptionId);
+		Task<IEnumerable<Snapshot>> GetAllAsync (long subscriptionId);
 	}
 }
